+++ packages/dashboard/src/components/Wizard/SetupWizardActionBar.types.ts
@@ -1,0 +1,24 @@
+import { IStyle } from 'office-ui-fabric-react/lib/Styling';
+import { IWizardStepAction, IWizardStepProps } from './Wizard.types';
+
+export interface ISetupWizardActionBarStyles {
+  root: IStyle;
+
+  actionLink: IStyle;
+
+  backAction: IStyle;
+
+  exitAction: IStyle;
+
+  mainAction: IStyle;
+}
+
+export interface ISetupWizardActionBarProps {
+  backClickAction: IWizardStepAction;
+
+  mainAction: IWizardStepAction;
+
+  exitWizardAction: IWizardStepAction;
+
+  currentStep: IWizardStepProps;
+}