+++ packages/office-ui-fabric-react/src/components/Callout/examples/Callout.Basic.Example.tsx
@@ -3,6 +3,7 @@ import { DefaultButton } from 'office-ui-fabric-react/lib/Button';
 import { Callout } from 'office-ui-fabric-react/lib/Callout';
 import { Link } from 'office-ui-fabric-react/lib/Link';
 import { getTheme, FontWeights, mergeStyleSets } from 'office-ui-fabric-react/lib/Styling';
+import { getId } from 'office-ui-fabric-react/lib/Utilities';
 
 export interface ICalloutBasicExampleState {
   isCalloutVisible?: boolean;
@@ -50,15 +51,15 @@ const styles = mergeStyleSets({
 
 // Example code
 export class CalloutBasicExample extends React.Component<{}, ICalloutBasicExampleState> {
+  public state: ICalloutBasicExampleState = {
+    isCalloutVisible: false
+  };
+
   private _menuButtonElement = React.createRef<HTMLDivElement>();
-
-  public constructor(props: {}) {
-    super(props);
-
-    this.state = {
-      isCalloutVisible: false
-    };
-  }
+  // Use getId() to ensure that the callout label and description IDs are unique on the page.
+  // (It's also okay to use plain strings without getId() and manually ensure their uniqueness.)
+  private _labelId: string = getId('callout-label');
+  private _descriptionId: string = getId('callout-description');
 
   public render(): JSX.Element {
     const { isCalloutVisible } = this.state;
@@ -66,13 +67,13 @@ export class CalloutBasicExample extends React.Component<{}, ICalloutBasicExampl
     return (
       <div>
         <div className={styles.buttonArea} ref={this._menuButtonElement}>
-          <DefaultButton id="toggleCallout" onClick={this._onShowMenuClicked} text={isCalloutVisible ? 'Hide callout' : 'Show callout'} />
+          <DefaultButton onClick={this._onShowMenuClicked} text={isCalloutVisible ? 'Hide callout' : 'Show callout'} />
         </div>
         <Callout
           className="ms-CalloutExample-callout"
-          ariaLabelledBy={'callout-label-1'}
-          ariaDescribedBy={'callout-description-1'}
-          role={'alertdialog'}
+          ariaLabelledBy={this._labelId}
+          ariaDescribedBy={this._descriptionId}
+          role="alertdialog"
           gapSpace={0}
           target={this._menuButtonElement.current}
           onDismiss={this._onCalloutDismiss}
@@ -80,13 +81,13 @@ export class CalloutBasicExample extends React.Component<{}, ICalloutBasicExampl
           hidden={!this.state.isCalloutVisible}
         >
           <div className="ms-CalloutExample-header">
-            <p className="ms-CalloutExample-title" id={'callout-label-1'}>
+            <p className="ms-CalloutExample-title" id={this._labelId}>
               All of your favorite people
             </p>
           </div>
           <div className="ms-CalloutExample-inner">
             <div className="ms-CalloutExample-content">
-              <p className="ms-CalloutExample-subText" id={'callout-description-1'}>
+              <p className="ms-CalloutExample-subText" id={this._descriptionId}>
                 Message body is optional. If help documentation is available, consider adding a link to learn more at the bottom.
               </p>
             </div>